adds proxy to vite config, fixes api routes, separates controllers and routers
