+++ client/index.html
@@ -9,10 +9,6 @@
   <body>
     <div id="modal"></div>
     <div id="root"></div>
-    <!-- <script type="module" src="/src/index.jsx"></script> -->
-
-    <!-- DEVELOPMENT -->
-    <script type="module" src="http://localhost:5173/@vite/client"></script>
-    <script type="module" src="http://localhost:5173/src/index.jsx"></script>
+    <script type="module" src="/src/main.jsx"></script>
   </body>
 </html>
