+++ client/vite.config.js
@@ -3,4 +3,12 @@ import react from "@vitejs/plugin-react";
 
 export default defineConfig({
   plugins: [react()],
+  server: {
+    proxy: {
+      // Proxying API requests
+      '/api': {
+        target: 'http://localhost:3000',
+      },
+    },
+  },
 });
