+++ compose.yml
@@ -28,7 +28,7 @@ services:
     depends_on:
       - server
     volumes:
-      - ./client:/client
+      - ./client:/app
       - /client/node_modules
 volumes:
   mongo-data:
