+++ server/config/posts.config.js
