+++ server/routes/auth.js
@@ -2,10 +2,10 @@ const express = require("express")
 const authRouter = new express.Router();
 const authController = require("../controllers/authController");
 
-// authRouter.get("/auth", (req, res) => {res.render("login")});
-authRouter.post("/auth", authController.handleLogin);
-authRouter.get("/auth/refresh", authController.handleRefreshToken)
-authRouter.post("/auth/logout", authController.handleLogout)
-authRouter.get("/auth/validatePath/:category/:postId?/:edit?", authController.validatePath);
+// authRouter.post("/api/auth/register", authController.handleNewUser);
+authRouter.post("/api/auth", authController.handleLogin);
+authRouter.get("/api/auth/refresh", authController.handleRefreshToken)
+authRouter.post("/api/auth/logout", authController.handleLogout)
+authRouter.get("/api/auth/validatePath/:category/:postId?/:edit?", authController.validatePath);
 
 module.exports = authRouter
