--- server/routes/register.js
@@ -1,7 +1,0 @@
-const express = require("express")
-const registerRouter = express.Router()
-const authController = require("../controllers/authController")
-
-registerRouter.post("/register", authController.handleNewUser);
-
-module.exports = registerRouter
