+++ server/routes/routers.js
@@ -1,8 +1,6 @@
 const authRouter = require("./auth")
 const postRouter = require("./post")
-const categoryRouter = require("./category")
-const registerRouter = require("./register")
 const tagRouter = require("./tag")
 
 
-module.exports = { authRouter, postRouter, registerRouter, categoryRouter, tagRouter }
+module.exports = { authRouter, postRouter, tagRouter }
